파일 분할 및 스왑 기능 구현
